# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para el catálogo de cartas coleccionables.",
        "title": "Trample API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Iniciar sesión",
                "parameters": [
                    {
                        "description": "Credenciales",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "token": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales inválidas",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Registrar usuario",
                "parameters": [
                    {
                        "description": "Datos del nuevo usuario",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "first_name": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "integer"
                                },
                                "last_name": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Email ya registrado",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Contraseña muy corta",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/scrydex/magic/cards": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic"
                ],
                "summary": "Buscar cartas Magic",
                "parameters": [
                    {
                        "description": "Filtros",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "expansion_code": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "rarity": {
                                    "type": "string"
                                },
                                "variants": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/catalog.Card"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/scrydex/magic/cards/{id}": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "magic"
                ],
                "summary": "Obtener carta Magic por ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID de la carta",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Variantes",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "variants": {
                                    "type": "array",
                                    "items": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/catalog.Card"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "catalog.Card": {
            "type": "object",
            "properties": {
                "abilities": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "artist": {
                    "type": "string"
                },
                "attacks": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "evolves_from": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "expansion": {
                    "$ref": "#/definitions/catalog.Expansion"
                },
                "expansion_sort_order": {
                    "type": "integer"
                },
                "external_id": {
                    "type": "string"
                },
                "flavor_text": {
                    "type": "string"
                },
                "hp": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.Image"
                    }
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "national_pokedex_numbers": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "number": {
                    "type": "string"
                },
                "printed_number": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                },
                "rarity_code": {
                    "type": "string"
                },
                "resistances": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "retreat_cost": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "subtypes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "supertype": {
                    "type": "string"
                },
                "types": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "variants": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.Variant"
                    }
                },
                "weaknesses": {
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
        "catalog.Expansion": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "language_code": {
                    "type": "string"
                },
                "logo": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "printed_total": {
                    "type": "integer"
                },
                "release_date": {
                    "type": "string"
                },
                "series": {
                    "type": "string"
                },
                "symbol": {
                    "type": "string"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "catalog.Image": {
            "type": "object",
            "properties": {
                "large": {
                    "type": "string"
                },
                "medium": {
                    "type": "string"
                },
                "small": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "catalog.Marketplace": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "purchase_url": {
                    "type": "string"
                }
            }
        },
        "catalog.Price": {
            "type": "object",
            "properties": {
                "low_cop": {
                    "type": "integer"
                },
                "low_usd": {
                    "type": "number"
                },
                "market_cop": {
                    "type": "integer"
                },
                "market_usd": {
                    "type": "number"
                },
                "trm_used": {
                    "type": "number"
                }
            }
        },
        "catalog.Variant": {
            "type": "object",
            "properties": {
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.Image"
                    }
                },
                "marketplaces": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/catalog.Marketplace"
                    }
                },
                "name": {
                    "type": "string"
                },
                "nm_price": {
                    "$ref": "#/definitions/catalog.Price"
                },
                "origin": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Ingresá el token con el prefijo Bearer. Ej: \"Bearer eyJ...\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}